Merged the files Falling_Oscilalations.py and ode_problems.py in to ode_problems.py
To Nate: Please push the code that gives the graphics that you posted on
the wiki.  If you want to put the velocity on there, we might as well
post the analytic velocity as well, which is given by just taking the
time derivative of the analytic solution.  I will add that code